refactor common module

+++ src/@app/common/country/application/delete/delete-countries.service.ts
@@ -1,10 +1,9 @@
+import { ICountryI18nRepository } from '../../domain/country-i18n.repository';
+import { ICountryRepository } from '../../domain/country.repository';
+import { AddCountriesContextEvent } from '../events/add-countries-context.event';
+import { CQMetadata, Operator, QueryStatement } from '@aurorajs.dev/core';
 import { Injectable } from '@nestjs/common';
 import { EventPublisher } from '@nestjs/cqrs';
-import { Operator, QueryStatement } from '@aurorajs.dev/core';
-import { CQMetadata } from '@aurorajs.dev/core';
-import { ICountryRepository } from '../../domain/country.repository';
-import { ICountryI18nRepository } from '../../domain/country-i18n.repository';
-import { AddCountriesContextEvent } from '../events/add-countries-context.event';
 
 @Injectable()
 export class DeleteCountriesService
@@ -21,6 +20,9 @@ export class DeleteCountriesService
         cQMetadata?: CQMetadata,
     ): Promise<void>
     {
+        const fallbackLang = cQMetadata.meta.fallbackLang;
+        const contentLanguage = cQMetadata.meta.contentLanguage;
+
         // get objects to delete
         const countries = await this.repository.get({
             queryStatement,
@@ -28,20 +30,43 @@ export class DeleteCountriesService
             cQMetadata,
         });
 
-        await this.repositoryI18n.delete({
-            queryStatement: {
-                where: {
-                    countryId: { [Operator.in]: countries.map(item => item.id) },
+        if (countries.length === 0) return;
+
+        if (countries[0].langId.value === fallbackLang.id)
+        {
+            // delete all translations if delete fallback language
+            await this.repository.delete({
+                queryStatement,
+                constraint,
+                cQMetadata,
+                deleteOptions: cQMetadata?.repositoryOptions,
+            });
+
+            await this.repositoryI18n.delete({
+                queryStatement: {
+                    where: {
+                        countryId: {
+                            [Operator.in]: countries.map(item => item.id),
+                        },
+                    },
                 },
-            },
-            deleteOptions: cQMetadata?.repositoryOptions,
-        });
-        await this.repository.delete({
-            queryStatement,
-            constraint,
-            cQMetadata,
-            deleteOptions: cQMetadata?.repositoryOptions,
-        });
+                deleteOptions: cQMetadata?.repositoryOptions,
+            });
+        }
+        else
+        {
+            await this.repositoryI18n.delete({
+                queryStatement: {
+                    where: {
+                        countryId: {
+                            [Operator.in]: countries.map(item => item.id),
+                        },
+                        langId: contentLanguage.id,
+                    },
+                },
+                deleteOptions: cQMetadata?.repositoryOptions,
+            });
+        }
 
         // create AddCountriesContextEvent to have object wrapper to add event publisher functionality
         // insert EventBus in object, to be able to apply and commit events
