+++ src/@app/common/lang/application/delete/delete-langs.service.ts
@@ -1,9 +1,8 @@
-import { Injectable } from '@nestjs/common';
-import { EventPublisher } from '@nestjs/cqrs';
-import { QueryStatement } from '@aurorajs.dev/core';
-import { CQMetadata } from '@aurorajs.dev/core';
 import { ILangRepository } from '../../domain/lang.repository';
 import { AddLangsContextEvent } from '../events/add-langs-context.event';
+import { CQMetadata, QueryStatement } from '@aurorajs.dev/core';
+import { Injectable } from '@nestjs/common';
+import { EventPublisher } from '@nestjs/cqrs';
 
 @Injectable()
 export class DeleteLangsService
@@ -25,6 +24,8 @@ export class DeleteLangsService
             constraint,
             cQMetadata,
         });
+
+        if (langs.length === 0) return;
 
         await this.repository.delete({
             queryStatement,
